feat: load data based on isLoaded flag
